Remove `api_nogui.py` since this is not necessary anymore
