+++ pyproject.toml
@@ -187,7 +187,6 @@ doctest_optionflags = [
 
 [tool.ruff]
 exclude = [
-    "api_nogui.py",
     "hyperspy/external/*",
     "examples",
     "conftest.py",
